+++ src/game.ts
@@ -451,31 +451,33 @@ export class Game {
             }
 
             sck.on("move", (move: string) => {
-                var points: Point[] = [];
-                switch (move) {
-                    case "right":
-                        points = this.currentShape.moveRight();
-                        break;
-                    case "left":
-                        points = this.currentShape.moveLeft();
-                        break;
-                    case "up":
-                        points = this.currentShape.rotate(true);
-                        break;
-                    case "down":
-                        points = this.currentShape.drop();
-                        break;
-                }
-                if (this.grid.isPosValid(points)) {
-                    if (move === "down") {
-                        clearTimeout(this.timerToken);
-                        this.timerToken = setInterval(() => {
-                            this.gameTimer();
-                        }, this.speed);
+                if (this.phase == Game.gameState.playing) {
+                    var points: Point[] = [];
+                    switch (move) {
+                        case "right":
+                            points = this.currentShape.moveRight();
+                            break;
+                        case "left":
+                            points = this.currentShape.moveLeft();
+                            break;
+                        case "up":
+                            points = this.currentShape.rotate(true);
+                            break;
+                        case "down":
+                            points = this.currentShape.drop();
+                            break;
                     }
+                    if (this.grid.isPosValid(points)) {
+                        if (move === "down") {
+                            clearTimeout(this.timerToken);
+                            this.timerToken = setInterval(() => {
+                                this.gameTimer();
+                            }, this.speed);
+                        }
 
-                    this.currentShape.setPos(points);
-                    io.emit("move", move);
+                        this.currentShape.setPos(points);
+                        io.emit("move", move);
+                    }
                 }
             });
 
@@ -489,10 +491,6 @@ export class Game {
 
             sck.on("increment level", () => {
                 this.incrementLevel();
-                clearTimeout(this.timerToken);
-                this.timerToken = setInterval(() => {
-                    this.gameTimer();
-                }, this.speed);
             });
         });
     }
@@ -505,7 +503,7 @@ export class Game {
         this.score = 0;
         this.rowsCompleted = 0;
         this.level = -1;
-        this.speed = 1000;
+        this.speed = 900;
         this.phase = Game.gameState.playing;
         this.randomShapes = [];
 
@@ -554,14 +552,15 @@ export class Game {
     }
 
     private incrementLevel() {
-        this.level++;
-        if (this.level < 10) {
-            this.speed = 1000 - (this.level * 100);
+        if (this.level < 7) {
+            this.level++;
+            this.speed -= 100;
+
+            clearTimeout(this.timerToken);
+            this.timerToken = setInterval((function (self) {
+                return function () { self.gameTimer(); };
+            })(this), this.speed);
         }
-        clearTimeout(this.timerToken);
-        this.timerToken = setInterval((function (self) {
-            return function () { self.gameTimer(); };
-        })(this), this.speed);
     }
 
     private togglePause() {
